requirements.txt
python-telegram-bot~=13.7
requests==2.32.3
python-dotenv==1.1.0
Pillow==10.3.0

